requirements/requirements.txt
amqp==2.6.0
asgiref==3.2.9
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-rest-hooks==1.5.0
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
ipaddress==1.0.23
itsdangerous==1.1.0
kombu==4.6.10
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
